Added get handlers for users and wrote separate tests.
