+++ api/v1/users/routes.js
@@ -3,41 +3,6 @@
 const Joi = require('joi');
 let internals = {};
 
-internals.getUsers = function (request, reply) {
-  if (request.query.name) {
-    return reply(internals.findUsers(request.query.name));
-  }
-  reply(internals.users);
-};
-
-internals.findUsers = function (name) {
-  return internals.users.filter((user) => {
-    return user.name.toLowerCase() === name.toLowerCase();
-  });
-};
-
-internals.getUser = function (request, reply) {
-  const filtered = internals.users.filter((user) => {
-    return user.id === parseInt(request.params.id);
-  }).pop();
-  reply(filtered);
-};
-
-internals.users = [
-  {
-    id: 1,
-    name: 'Vishnu'
-  },
-  {
-    id: 2,
-    name: 'Vasanth'
-  },
-  {
-    id: 3,
-    name: 'Venkat'
-  }
-];
-
 module.exports = [
   {
     method: 'GET',
@@ -48,13 +13,13 @@ module.exports = [
           name: Joi.string()
         }
       },
-      handler: internals.getUsers
+      handler: require('./getByName.js')
     }
   },
   {
     method: 'GET',
     path: '/api/v1/users/{id}',
-    handler: internals.getUser
+    handler: require('./getById.js')
   },
   {
     method: 'POST',
