+++ test/api/v1/users/post.js
@@ -8,7 +8,7 @@ const describe = lab.describe;
 const it = lab.it;
 const expect = Code.expect;
 
-const server = require('../../../server');
+const server = require('../../../../server');
 
 describe('/api/v1/users endpoints', () => {
   it('POST /api/v1/users', (done) => {
@@ -24,7 +24,6 @@ describe('/api/v1/users endpoints', () => {
     };
     server.inject(request, (res) => {
       expect(res.statusCode).to.equal(200);
-      console.log('TES : '+ res.result)
       expect(res.result.username).to.equal(user.username);
       expect(res.result.password).to.equal(user.password);
       expect(res.result.email).to.equal(user.email);
